% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dist_matrix_function.R
\name{dist_matrix_function}
\alias{dist_matrix_function}
\title{Creates a matrix with the distances between the different N points}
\usage{
dist_matrix_function(dataset)
}
\arguments{
\item{dataset}{dataset is a matrix of NxM points}
}
\value{
distance_matrix_NxN distance_matrix_NxN is a NxN matrix
which contains the euclidean distance between all points of our dataset
}
\description{
Creates a matrix with the distances between the different N points
}
\examples{
dataset <- matrix(c(1,1,2,2,3,3,4,4,5,5),5,2, byrow = TRUE)
dist_matrix_function(dataset)
}

